leads_parser/serializer.py
from .models import Customer
from rest_framework.serializers import ModelSerializer
from rest_framework import serializers


class CustomerSerializer(ModelSerializer):
    class Meta:
        model = Customer
        fields = "__all__"


class FileUploadSerializer(serializers.Serializer):
    file = serializers.FileField()

    class Meta:
        fields = ('file',)
